Factored svoteban into the common implementation
