+++ lib/commands/implementations/voteban.js
@@ -1,3 +1,4 @@
+const _ = require('lodash');
 const moment = require('moment');
 
 const Command = require('../command-interface');
@@ -7,7 +8,14 @@ const basePunishmentHelper = require('../base-punishment-helper');
 const formatDuration = require('../../chat-utils/format-duration');
 
 let voteBanStarted = false;
-function voteBan(ipBan, defaultBanTime) {
+const weightedTranslateMap = {
+  flair8: 16,
+  flair3: 8,
+  flair1: 4,
+  flair13: 2,
+  flair9: 2,
+};
+function voteBan(ipBan, defaultBanTime, weighted) {
   return function ban(input, services, rawMessage) {
     if (voteBanStarted) {
       return new CommandOutput(
@@ -41,11 +49,21 @@ function voteBan(ipBan, defaultBanTime) {
       if (votedMap[data.user]) {
         return;
       }
+      let votes = 1;
+      if (weighted) {
+        votes = _.max(
+          Object.keys(weightedTranslateMap).map(k => {
+            const idx = data.roles.indexOf(k);
+            if (idx === -1) return 1;
+            return weightedTranslateMap[k];
+          }),
+        );
+      }
       const message = data.message.trim();
       if (message.trim() === '1' || message.trim() === '2') {
         votedMap[data.user] = true;
         // eslint-disable-next-line no-unused-expressions
-        message.trim() === '1' ? (yes += 1) : (no += 1);
+        message.trim() === '1' ? (yes += votes) : (no += votes);
       }
     });
 
@@ -85,14 +103,28 @@ function voteBan(ipBan, defaultBanTime) {
 
 module.exports = {
   voteIpban: new Command(
-    voteBan(true, 'perm'),
+    voteBan(true, 'perm', false),
     false,
     true,
     /((?:\d+[HMDSWwhmds])|(?:perm))?\s?(\w+)(?:\s(.*))?/,
     null,
   ),
   voteBan: new Command(
-    voteBan(false, 600),
+    voteBan(false, 600, false),
+    false,
+    true,
+    /((?:\d+[HMDSWwhmds])|(?:perm))?\s?(\w+)(?:\s(.*))?/,
+    null,
+  ),
+  svoteIpban: new Command(
+    voteBan(true, 'perm', true),
+    false,
+    true,
+    /((?:\d+[HMDSWwhmds])|(?:perm))?\s?(\w+)(?:\s(.*))?/,
+    null,
+  ),
+  svoteBan: new Command(
+    voteBan(false, 600, true),
     false,
     true,
     /((?:\d+[HMDSWwhmds])|(?:perm))?\s?(\w+)(?:\s(.*))?/,
