+++ lib/configuration/configure-commands.js
@@ -27,8 +27,7 @@ const {
   getThreadInfo,
 } = require('../commands/implementations/reddit-questions');
 const { getDuo, updateDuo } = require('../commands/implementations/duo');
-const { voteBan, voteIpban } = require('../commands/implementations/voteban');
-const { svoteBan, svoteIpban } = require('../commands/implementations/voteban');
+const { voteBan, voteIpban, svoteBan, svoteIpban } = require('../commands/implementations/voteban');
 const { setDeaths, getDeaths, incrementDeaths } = require('../commands/implementations/deathcount');
 const { gulag } = require('../commands/implementations/gulag');
